Prevent Deducting Price When Park Has No Money
Although scenarios can specify "No money" as a park flag, the park still
has a resovoir of cash that this plugin will deduct from. In cases like
this, don't deduct any cash from the park. This prevents issues where
the park has $0 to start out with, but is set to be an infinite money
scenario and thus the plugin does not allow building benches/bins
immediately.

Fixes #31