fix(types): resolve type checks after queue lines fix
Since #55 was fixed, there has been a type error in the compiler, preventing
builds and dependency updates. My fault for not noticing this, but this should
effectively be the same thing without the type issues since it's not relying
on the whole bitwise and/or thing to cast types in the runtime.