Fix Bin Number Change in OpenRCT2
As a result of openrct2/openrct2#20177, we no longer need to mess with the
index number of the selected bin or bench. Adding the `+ 1` here was breaking
functionality on newer versions of OpenRCT2, and its removal has been confirmed
to work, so we're codifying it in the plugin now.

Shoutouts to @KyuuGryphon, @Soma91, and @JensMakerAdventures for finding the
fix for this one! Much obliged.

Fixes #76